backend/pyproject.toml
[tool.poetry]
name = "gen-ui-backend"
version = "0.1.0"
description = ""
authors = ["shreyasskasetty <shreyas30kasetty@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.12,>=3.9.0"
langchain-core = "0.3.15"
typer = ">=0.9.0,<0.10.0"
langchain = "0.3.6"
langchain-openai = "^0.2.4"
pymongo = "^4.6.3"
python-dotenv = "^1.0.1"
motor = "^3.4.0"
langgraph = "0.2.43"
langserve = "^0.3.0"
fastapi = ">=0.110.2,<1"
uvicorn = ">=0.23.2,<0.24.0"
pydantic = "^2.9.2"
rich = "^13.7.1"
langchain-community = "0.3.4"
unstructured = {extras = ["all-docs"], version = "^0.13.4"}
langgraph-cli = "^0.1.46"
langchain-ollama = "^0.2.0"
sse-starlette = "^2.1.3"
composio-langgraph = "^0.5.37"
tiktoken = "^0.8.0"
langchainhub = "0.1.21"
chromadb = "^0.5.16"
langchain-text-splitters = "^0.3.1"
pypdf2 = "^3.0.1"

[tool.poetry.scripts]
start = "gen_ui_backend.server:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
